Adding pause and resume methods
Consider a scenario where annyang is in  listening mode and is executing a command callback function, and within the command we are using speechsynthesis to speak out a particular text, if the text in the speach matches with one of the registered annyang command, it would cause that command to be executed. See the code below

Ideally, it would be nice if we could pause annyang when speachSynthesis is active mode and resume it when the speachSynthesis ends as shown below

This PR implements a simple solution to achieve this. Please suggest if there is a better way of doing this.