Added support for Regular Expression commands
### Added
You can now pass an object containing a RegExp and a callback function
to the addCommands() or init() methods. See
https://github.com/TalAter/annyang/tree/master/docs#commands-object for
details
### Changed
In debug mode, adding commands no longer logs a message with the number
of commands loaded. Instead it logs individual messages for each
command registered along with it's name.